박지훈 : 제품별 매출 통계 완료, 지역별 매출 통계 95% 진행 완료 get방식 진행중 04/29 16:38

+++ src/main/java/com/naedam/mir9/statistics/model/vo/ProductStatisticVo.java
@@ -20,7 +20,7 @@ public class ProductStatisticVo implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
 	private String paidAt;
-	private double amount;
+	private int amount;
 	private String modelName;
 	private String productName;
 	private String regStartdate;
@@ -33,6 +33,6 @@ public class ProductStatisticVo implements Serializable {
 	private int parentNo;
 	private int paymentCount;
 	private int returnCount;
-	private double paymentSum;
+	private int paymentSum;
 	private int productRank;
 }
